requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
huggingface-hub
sentence-transformers
python-dotenv
faiss-cpu
